feat: Move enums to a file and Support exporting to YAML
